requirements.txt
wandb
jiwer
torch
torchaudio
torchvision 
transformers
librosa
soundfile
hydra-core
datasets
numpy=1.26.4